fix(render): a mesh over 65,535 vertices WRAPPED its indices — the head lost its face
`MeshData.indices` was `List[UInt16]`, so an index past 65,535 wrapped instead
of failing. The triangles beyond the wrap are still DRAWN, connecting the
wrong vertices, so a large mesh renders partially and nothing raises.

Menagerie's ToddlerBot head is 62,912 triangles = **188,736 vertices**.
Measured on the old code: **123,200 of 188,736 indices wrong** — 65% of the
mesh pointing somewhere else. On screen that was a robot with no face and one
eye.

## Why it took a person looking at a screen to find

Every check upstream passed while the picture was wrong. The `<mesh>` asset
resolved, the STL loaded and uploaded without raising, `geom_mesh_id` was
valid, the body sat at MuJoCo's exact position (all 46 verified against
`mj_forward`), and our visible geom set matched MuJoCo's element for element —
45 geoms plus the plane. The loss happens INSIDE one mesh, downstream of
everything a parse-level or kinematics-level test can see.

I had already checked the asset table, the file loads, the draw skips and the
body poses, and concluded the render was complete. It was, at the granularity
I was measuring. "The head has one eye" was the only signal that pointed
below it.

## The fix

`indices` is `List[UInt32]`, `index_byte_size` is `* 4`, all three draw calls
pass `GPU_INDEXELEMENTSIZE_32BIT`, and every producer (`gpu_mesh`'s
primitives, `stl_loader`, the skin seed) emits `UInt32`.

⚠ OUR OWN ASSETS ARE ALL UNDER THE LIMIT — it takes a ~3 MB STL to reach it —
which is why this survived. So `tests/render/test_mesh_index_width.mojo` uses
an EXTERNAL fixture deliberately: a gate built from this tree's own meshes
could not fail. It asserts the fixture EXCEEDS 65,535 first (a smaller one
cannot tell the two element sizes apart, so every other arm would pass on the
broken code), then that the largest index addresses the last vertex rather
than 65,535, and that no index is wrong at all — an STL shares no vertices, so
index i must equal i, which catches the FIRST wrap and not just the maximum.
A generated sphere covers the small-mesh path, since widening the element size
changes every draw call and not only the large ones.

Verified by stashing the fix: 3/5 with 123,200 wraps reported. 5/5 with it.

Regressions: `test_dog_skin` 3/3 (skinning shares this buffer), ray-pick
35/35, `dog_skin_view` builds, and so_arm101 / quadruped / sawyer render with
zero mesh warnings.

+++ mojo_rl/render/stl_loader.mojo
@@ -79,7 +79,7 @@ def load_stl(path: String) raises -> MeshData:
             mesh.vertices.append(
                 GPUVertex(px=px, py=py, pz=pz, nx=nx, ny=ny, nz=nz)
             )
-            mesh.indices.append(UInt16(tri * 3 + v))
+            mesh.indices.append(UInt32(tri * 3 + v))
 
         offset += 50
